+++ src/index.ts
@@ -1,6 +1,7 @@
+import { BurgerButton } from './components/BurgerButton/BurgerButton'
 import { Button } from './components/Button/Button'
 import { Confirm } from './components/Confirm/Confirm'
 import { Date } from './components/Date/Date'
 import { Modal } from './components/Modal/Modal'
 
-export { Button, Confirm, Date, Modal }
+export { BurgerButton, Button, Confirm, Date, Modal }