+++ CHANGELOG.md
@@ -1,10 +1,15 @@
+## Version 0.4.0
+
+* Add **BurgerButton** component
+* Confirm and Modal **close** prop is changed to **onClose**
+
 ## Version 0.3.0
 
 * Set standard styles (color palette)
 * Add customization option (it is possible to set **className** as property to all the components to add custom styles)
 * Button **Dangerous** flag is replaced by **Outlined**. The styles are also changed
-* Confirm **children** prop changed to **title**. View is not changed
-* Date **country** prop changed to region
+* Confirm **children** prop is changed to **title**. View is not changed
+* Date **country** prop is changed to region
 
 ## Version 0.2.0